+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/RobotKeywordCall.java
@@ -24,4 +24,6 @@ public interface RobotKeywordCall extends PsiNameIdentifierOwner, RobotKeywordCa
 
   @NotNull RobotKeywordCallName getNameIdentifier();
 
+  @NotNull String getSimpleKeywordName();
+
 }