+++ src/main/gen/dev/xeonkryptos/xeonrobotframeworkplugin/psi/element/impl/RobotKeywordCallImpl.java
@@ -61,4 +61,9 @@ public class RobotKeywordCallImpl extends RobotKeywordCallExtension implements R
     return RobotPsiImplUtil.getNameIdentifier(this);
   }
 
+  @Override
+  public @NotNull String getSimpleKeywordName() {
+    return RobotPsiImplUtil.getSimpleKeywordName(this);
+  }
+
 }